% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{jsonld}
\alias{jsonld}
\alias{compact}
\alias{expand}
\alias{flatten}
\alias{frame}
\alias{from_rdf}
\alias{to_rdf}
\alias{normalize}
\title{Working with JSON-LD}
\usage{
compact(doc, context)

expand(compacted)

flatten(doc)

frame(doc)

from_rdf(rdf, format = "application/nquads")

to_rdf(doc, format = "application/nquads")

normalize(doc, algorithm = "URDNA2015", format = "application/nquads")
}
\arguments{
\item{doc}{string with JSON-LD document}

\item{context}{string with JSON-LD context}

\item{compacted}{string with JSON message}

\item{rdf}{string with RDF text}

\item{format}{a MIME type}

\item{algorithm}{normalization algorithm}
}
\description{
Wrappers for converting, expanding and compating JSON-LD documents. All parameters
and return values must be JSON strings. Use \link[jsonlite:toJSON]{jsonlite::toJSON} and \link[jsonlite:fromJSON]{jsonlite::fromJSON}
to convert between R objects and JSON format. The \href{https://github.com/ropensci/jsonld#readme}{readme}
has basic examples.
}

